[#1] :recycle: refactor(nav-bar): icon, description pair 상수 파일 분리
