+++ client/src/components/common/NavBar.tsx
@@ -4,42 +4,12 @@ import { NavLink } from "react-router-dom";
 import { useRecoilValue } from "recoil";
 
 import { activeMenuAtom } from "../../recoil/nav-bar/atom";
+import { ICON_DESC_PAIR_OBJ } from "src/constants/ICON_DESC_PAIR_OBJ";
 import { PRIVATE_ROUTE } from "src/router/ROUTE_INFO";
-
-import ColoredDiaryImg from "../../assets/icons/colored_note.png";
-import ColoredGrapeImg from "../../assets/icons/colored_grape.png";
-import ColoredMyPageImg from "../../assets/icons/colored_mypage.png";
-import GrayDiaryImg from "../../assets/icons/gray_note.png";
-import GrayGrapeImg from "../../assets/icons/gray_grape.png";
-import GrayMyPageImg from "../../assets/icons/gray_mypage.png";
-
-interface PairObj {
-  readonly [key: string]: IconDescPair;
-}
-
-interface IconDescPair {
-  readonly icon: [string, string];
-  readonly description: string;
-}
 
 interface Menu {
   readonly menu: string;
 }
-
-const ICON_DESC_PAIR_OBJ: PairObj = {
-  diary: {
-    icon: [ColoredDiaryImg, GrayDiaryImg],
-    description: "공유일기",
-  },
-  grape: {
-    icon: [ColoredGrapeImg, GrayGrapeImg],
-    description: "포도알",
-  },
-  myPage: {
-    icon: [ColoredMyPageImg, GrayMyPageImg],
-    description: "마이페이지",
-  },
-};
 
 const BarContainer = tw.div`
   flex justify-around px-12 mx-auto w-10/12 h-[80px] bg-white/60 rounded-2xl fixed left-1/2 -translate-x-1/2 bottom-10
